# entrypoints/api/openapi.yml
openapi: 3.0.3
info:
  title: omnius-opxs-api
  description: ''
  contact:
    name: Lyrise
    email: lyrise1984@gmail.com
  license:
    name: MIT
  version: 0.1.0
servers:
- url: https://localhost.omnius-labs.com/api
  description: local
paths:
  /api/health:
    get:
      tags:
      - health
      operationId: health
      responses:
        '200':
          description: ''
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorMessage'
  /api/v1/auth/email/login:
    post:
      tags:
      - auth
      operationId: authEmailLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthToken'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorMessage'
  /api/v1/auth/email/register:
    post:
      tags:
      - auth
      operationId: authEmailRegister
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterInput'
        required: true
      responses:
        '200':
          description: ''
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorMessage'
  /api/v1/auth/google/login:
    post:
      tags:
      - auth
      operationId: authGoogleLogin
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthToken'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorMessage'
  /api/v1/auth/google/nonce:
    get:
      tags:
      - auth
      operationId: authGoogleNonce
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NonceOutput'
  /api/v1/auth/google/register:
    post:
      tags:
      - auth
      operationId: authGoogleRegister
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthToken'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorMessage'
  /api/v1/auth/me:
    get:
      tags:
      - auth
      operationId: authMe
      responses:
        '200':
          description: ''
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorMessage'
      security:
      - bearer_token: []
  /api/v1/file-convert/image/status:
    get:
      tags:
      - file-convert
      operationId: fileConvertImageStatus
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusInput'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusOutput'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorMessage'
      security:
      - bearer_token: []
  /api/v1/file-convert/image/upload:
    post:
      tags:
      - file-convert
      operationId: fileConvertImageUpload
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadInput'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadOutput'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorMessage'
      security:
      - bearer_token: []
components:
  schemas:
    ApiErrorCode:
      type: string
      enum:
      - InternalServerError
      - InvalidRequest
      - NotFound
      - TokenExpired
      - Unauthorized
      - Duplicated
    ApiErrorMessage:
      type: object
      required:
      - error_code
      properties:
        error_code:
          $ref: '#/components/schemas/ApiErrorCode'
    AuthToken:
      type: object
      required:
      - access_token
      - access_token_expires_at
      - refresh_token
      - refresh_token_expires_at
      properties:
        access_token:
          type: string
        access_token_expires_at:
          type: string
          format: date-time
        refresh_token:
          type: string
        refresh_token_expires_at:
          type: string
          format: date-time
    FileConvertImageInputFileType:
      type: string
      enum:
      - unknown
      - gif
      - jpg
      - png
      - webp
      - bmp
      - heif
      - heic
      - avif
      - svg
    FileConvertImageOutputFileType:
      type: string
      enum:
      - gif
      - jpg
      - png
      - webp
      - bmp
      - avif
      - svg
    FileConvertJobStatus:
      type: string
      enum:
      - Unknown
      - Preparing
      - Waiting
      - Processing
      - Completed
      - Rejected
      - Failed
    LoginInput:
      type: object
      required:
      - redirect_uri
      - code
      properties:
        code:
          type: string
        redirect_uri:
          type: string
    NonceOutput:
      type: object
      required:
      - value
      properties:
        value:
          type: string
    RegisterInput:
      type: object
      required:
      - redirect_uri
      - code
      properties:
        code:
          type: string
        redirect_uri:
          type: string
    StatusInput:
      type: object
      required:
      - job_id
      properties:
        job_id:
          type: string
    StatusOutput:
      type: object
      required:
      - status
      properties:
        download_url:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/FileConvertJobStatus'
    UploadInput:
      type: object
      required:
      - in_file_name
      - in_type
      - out_file_name
      - out_type
      properties:
        in_file_name:
          type: string
        in_type:
          $ref: '#/components/schemas/FileConvertImageInputFileType'
        out_file_name:
          type: string
        out_type:
          $ref: '#/components/schemas/FileConvertImageOutputFileType'
    UploadOutput:
      type: object
      required:
      - job_id
      - upload_url
      properties:
        job_id:
          type: string
        upload_url:
          type: string
  securitySchemes:
    bearer_token:
      type: http
      scheme: bearer